web/backend/requirements.txt
fastapi>=0.120.0
uvicorn[standard]>=0.38.0
python-multipart>=0.0.20
python-dotenv>=1.2.0
google-auth>=2.42.0
google-auth-httplib2>=0.2.1
requests>=2.32.0
google-cloud-firestore>=2.13.0
google-cloud-run>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0

